requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=2.0.0=pypi_0
ale-py=0.7.4=pypi_0
asttokens=2.4.1=pyhd8ed1ab_0
autorom=0.6.1=pypi_0
autorom-accept-rom-license=0.6.1=pypi_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=pyhd8ed1ab_3
backports.functools_lru_cache=1.6.5=pyhd8ed1ab_0
blas=2.119=mkl
blas-devel=3.9.0=19_win64_mkl
brotli-python=1.1.0=py38hd3f51b4_1
bzip2=1.0.8=hcfcfb64_5
ca-certificates=2023.7.22=h56e8100_0
cachetools=5.3.2=pypi_0
certifi=2023.7.22=pypi_0
charset-normalizer=3.3.2=pyhd8ed1ab_0
click=8.1.7=pypi_0
cloudpickle=3.0.0=pypi_0
colorama=0.4.6=pyhd8ed1ab_0
comm=0.1.4=pyhd8ed1ab_0
contourpy=1.1.1=pypi_0
cuda-cccl=12.3.52=0
cuda-cudart=12.1.105=0
cuda-cudart-dev=12.1.105=0
cuda-cupti=12.1.105=0
cuda-libraries=12.1.0=0
cuda-libraries-dev=12.1.0=0
cuda-nvrtc=12.1.105=0
cuda-nvrtc-dev=12.1.105=0
cuda-nvtx=12.1.105=0
cuda-opencl=12.3.52=0
cuda-opencl-dev=12.3.52=0
cuda-profiler-api=12.3.52=0
cuda-runtime=12.1.0=0
cycler=0.12.1=pypi_0
debugpy=1.8.0=py38hd3f51b4_1
decorator=5.1.1=pyhd8ed1ab_0
executing=2.0.1=pyhd8ed1ab_0
filelock=3.13.1=pyhd8ed1ab_0
fonttools=4.44.0=pypi_0
freetype=2.12.1=hdaf720e_2
fsspec=2023.10.0=pypi_0
google-auth=2.23.4=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
grpcio=1.59.2=pypi_0
gym=0.21.0=pypi_0
gym-retro=0.8.0=pypi_0
gym-super-mario-bros=7.3.0=pypi_0
idna=3.4=pyhd8ed1ab_0
importlib-metadata=4.13.0=pypi_0
importlib-resources=6.1.1=pypi_0
importlib_metadata=6.8.0=hd8ed1ab_0
intel-openmp=2023.2.0=h57928b3_50497
ipykernel=6.26.0=pyha63f2e9_0
ipython=8.12.0=pyh08f2357_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.2=pyhd8ed1ab_1
jupyter_client=8.6.0=pyhd8ed1ab_0
jupyter_core=5.5.0=py38haa244fe_0
kiwisolver=1.4.5=pypi_0
lcms2=2.15=h67d730c_3
lerc=4.0.0=h63175ca_0
libblas=3.9.0=19_win64_mkl
libcblas=3.9.0=19_win64_mkl
libcublas=12.1.0.26=0
libcublas-dev=12.1.0.26=0
libcufft=11.0.2.4=0
libcufft-dev=11.0.2.4=0
libcurand=10.3.4.52=0
libcurand-dev=10.3.4.52=0
libcusolver=11.4.4.55=0
libcusolver-dev=11.4.4.55=0
libcusparse=12.0.2.55=0
libcusparse-dev=12.0.2.55=0
libdeflate=1.19=hcfcfb64_0
libffi=3.4.2=h8ffe710_5
libhwloc=2.9.3=default_haede6df_1009
libiconv=1.17=h8ffe710_0
libjpeg-turbo=3.0.0=hcfcfb64_1
liblapack=3.9.0=19_win64_mkl
liblapacke=3.9.0=19_win64_mkl
libnpp=12.0.2.50=0
libnpp-dev=12.0.2.50=0
libnvjitlink=12.1.105=0
libnvjitlink-dev=12.1.105=0
libnvjpeg=12.1.1.14=0
libnvjpeg-dev=12.1.1.14=0
libpng=1.6.39=h19919ed_0
libsodium=1.0.18=h8d14728_1
libsqlite=3.44.0=hcfcfb64_0
libtiff=4.6.0=h6e2ebb7_2
libuv=1.44.2=hcfcfb64_1
libwebp-base=1.3.2=hcfcfb64_0
libxcb=1.15=hcd874cb_0
libxml2=2.11.5=hc3477c8_1
libzlib=1.2.13=hcfcfb64_5
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markdown=3.5.1=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.3=py38h91455d4_1
matplotlib=3.7.3=pypi_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mdurl=0.1.2=pypi_0
mkl=2023.2.0=h6a75c08_50496
mkl-devel=2023.2.0=h57928b3_50496
mkl-include=2023.2.0=h6a75c08_50496
mpmath=1.3.0=pyhd8ed1ab_0
msys2-conda-epoch=20160418=1
nes-py=8.2.1=pypi_0
nest-asyncio=1.5.8=pyhd8ed1ab_0
networkx=3.1=pypi_0
numpy=1.24.4=py38h1d91fd2_0
oauthlib=3.2.2=pypi_0
opencv-python=4.8.1.78=pypi_0
openjpeg=2.5.0=h3d672ee_3
openssl=3.1.4=hcfcfb64_0
packaging=23.2=pyhd8ed1ab_0
pandas=2.0.3=pypi_0
parso=0.8.3=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=10.1.0=py38hc375fad_0
pip=23.3.1=pyhd8ed1ab_0
platformdirs=3.11.0=pyhd8ed1ab_0
prompt-toolkit=3.0.39=pyha770c72_0
prompt_toolkit=3.0.39=hd8ed1ab_0
protobuf=4.25.0=pypi_0
psutil=5.9.6=pypi_0
pthread-stubs=0.4=hcd874cb_1001
pthreads-win32=2.9.1=hfa6e2cd_3
pure_eval=0.2.2=pyhd8ed1ab_0
pyasn1=0.5.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
pyglet=1.5.21=pypi_0
pygments=2.16.1=pyhd8ed1ab_0
pyparsing=3.1.1=pypi_0
pysocks=1.7.1=pyh0701188_6
python=3.8.18=h4de0772_0_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.8=4_cp38
pytorch=2.1.0=py3.8_cuda12.1_cudnn8_0
pytorch-cuda=12.1=hde6ce7c_5
pytorch-mutex=1.0=cuda
pytz=2023.3.post1=pypi_0
pywin32=306=py38hd3f51b4_2
pyyaml=6.0.1=py38h91455d4_1
pyzmq=25.1.1=py38hdecc13b_2
requests=2.31.0=pyhd8ed1ab_0
requests-oauthlib=1.3.1=pypi_0
rich=13.6.0=pypi_0
rsa=4.9=pypi_0
setuptools=65.5.0=pypi_0
six=1.16.0=pyh6c4a22f_0
stable-baselines3=1.8.0=pypi_0
stack_data=0.6.2=pyhd8ed1ab_0
sympy=1.12=pyh04b8f61_3
tbb=2021.10.0=h91493d7_2
tensorboard=2.14.0=pypi_0
tensorboard-data-server=0.7.2=pypi_0
tk=8.6.13=h5226925_1
torch=2.1.0=pypi_0
torchaudio=2.1.0=pypi_0
torchvision=0.16.0=pypi_0
tornado=6.3.3=py38h91455d4_1
tqdm=4.66.1=pypi_0
traitlets=5.13.0=pyhd8ed1ab_0
typing-extensions=4.8.0=hd8ed1ab_0
typing_extensions=4.8.0=pyha770c72_0
tzdata=2023.3=pypi_0
ucrt=10.0.22621.0=h57928b3_0
urllib3=2.0.7=pyhd8ed1ab_0
vc=14.3=h64f974e_17
vc14_runtime=14.36.32532=hdcecf7f_17
vs2015_runtime=14.36.32532=h05e6639_17
wcwidth=0.2.9=pyhd8ed1ab_0
werkzeug=3.0.1=pypi_0
wheel=0.38.4=pypi_0
win_inet_pton=1.1.0=pyhd8ed1ab_6
xorg-libxau=1.0.11=hcd874cb_0
xorg-libxdmcp=1.1.3=hcd874cb_0
xz=5.2.6=h8d14728_0
yaml=0.2.5=h8ffe710_2
zeromq=4.3.5=h63175ca_0
zipp=3.17.0=pyhd8ed1ab_0
zstd=1.5.5=h12be248_0
